requirements.txt
python-telegram-bot
google-generativeai
python-dotenv
feedparser
edge-tts
requests
